+++ plugins/backend/backstage-plugin-argo-cd-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @roadiehq/backstage-plugin-argo-cd-backend
 
+## 2.5.3
+
+### Patch Changes
+
+- eaa0bb2: update dependencies
+
+## 2.5.2
+
+### Patch Changes
+
+- 231f37e: Added logging to improve visibility
+
+## 2.5.1
+
+### Patch Changes
+
+- e017302: 1. Get Application App Data has a changed thrown error message. 2. Failing on a wider spectrum of error messages when deleting projects. If you were expecting a false when the project failed to delete for these other possible reasons, now the function will throw for those other possible reasons. 3. Logging thrown errors for the endpoints that create an argo configuration, or delete an argo configuration. 4. Added unit tests.
+
 ## 2.5.0
 
 ### Minor Changes